+++ src/main/java/tk/p4rty/cachemyenderchest/mixin/ClientPlayNetworkHandlerMixin.java
@@ -15,7 +15,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import tk.p4rty.cachemyenderchest.Main;
+import tk.p4rty.cachemyenderchest.CacheMyEnderchest;
 
 @Mixin(ClientPlayNetworkHandler.class)
 public abstract class ClientPlayNetworkHandlerMixin implements ClientPlayPacketListener {
@@ -28,12 +28,12 @@ public abstract class ClientPlayNetworkHandlerMixin implements ClientPlayPacketL
             for (int i = 0; i < packet.getContents().size() - 36; i++) {
                 items[i] = packet.getContents().get(i);
             }
-            Main.ecinv = new SimpleInventory(items);
+            CacheMyEnderchest.ecinv = new SimpleInventory(items);
         }
     }
     @Inject(method = "onGameJoin", at = @At("TAIL"))
     private void onOnGameJoin(GameJoinS2CPacket packet, CallbackInfo ci) {
-        Main.ecinv = new EnderChestInventory();
+        CacheMyEnderchest.ecinv = new EnderChestInventory();
     }
 
 }
