+++ src/main/java/tk/p4rty/cachemyenderchest/mixin/ScreenMixin.java
@@ -13,7 +13,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import tk.p4rty.cachemyenderchest.Main;
+import tk.p4rty.cachemyenderchest.CacheMyEnderchest;
 
 @Mixin(Screen.class)
 public abstract class ScreenMixin extends AbstractParentElement implements Drawable {
@@ -21,9 +21,10 @@ public abstract class ScreenMixin extends AbstractParentElement implements Drawa
 
     @Inject(method = "keyPressed", at = @At("HEAD"))
     public void onKeyPressed(int keyCode, int scanCode, int modifiers, CallbackInfoReturnable<Boolean> cir) {
+        if (this.client == null) return;
         if ((this.client.currentScreen instanceof GenericContainerScreen) && this.client.currentScreen.getTitle().equals(Text.translatable("container.enderchest"))) {
             if (keyCode == 256 || keyCode == 69) {
-                Main.ecinv = (SimpleInventory) ((GenericContainerScreen) this.client.currentScreen).getScreenHandler().getInventory();
+                CacheMyEnderchest.ecinv = (SimpleInventory) ((GenericContainerScreen) this.client.currentScreen).getScreenHandler().getInventory();
             }
         }
     }
